backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autoflake==2.3.1
autopep8==2.3.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.15
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mysqlclient==2.2.5
packaging==24.2
pillow==10.2.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyGithub==2.5.0
PyJWT==2.9.0
PyNaCl==1.5.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
setuptools==75.1.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.44.0
wrapt==1.16.0
